Add support for graphical banners
Add the ability to write a graphical banner (in PNG format) to the terminal, if the console is detected to support graphics. The image is streamed from a given input stream. Optional values can be given to specify how many character cells (width and height) should be occupied by the image.